src/requirements.txt
mediapipe==0.10.18
numpy==1.24.2
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python==4.12.0.88
picamera2==0.3.31
pyserial==3.5
